Reorganize code into `scenes` and `components`
