--- src/App.js
@@ -1,8 +1,0 @@
-import React from "react";
-import "./App.css";
-
-import Game from "./Game";
-
-const App = () => <Game />;
-
-export default App;
