+++ src/components/FeedbackMessage.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+const Message = styled.h2`
+  margin: 0;
+`;
+
+const FeedbackMessage = styled(Message)`
+  font-size: 2.3em;
+  color: white;
+`;
+
+export { Message, FeedbackMessage };
